+++ CHANGELOG.md
@@ -1,5 +1,6 @@
 ## 3.2.0
 - Bug fix for marshal of structs in Opal
+- Add last_input and last_prompt at start of take
 
 ## 3.1.0 - April 8, 2024
 - Dispatcher prioritizes strict token matches